+++ jim_hats_mobile/lib/presentation/views/gym_challenge_details/gym_challenge_details_page.dart
@@ -32,7 +32,7 @@ class GymChallengeDetailsView extends StatelessWidget {
       body: SafeArea(
         child: Padding(
           padding: EdgeInsets.symmetric(
-            horizontal: AppSpacings.horizontalPadding.toDouble(),
+            horizontal: AppSpacings.horizontalPadding,
           ),
           child: BlocBuilder<GymChallengeDetailsPageCubit,
               GymChallengeDetailsPageState>(
