+++ jim_hats_mobile/lib/presentation/views/join_group/join_group_page.dart
@@ -33,7 +33,7 @@ class _JoinGroupViewState extends State<JoinGroupView> {
       body: SafeArea(
         child: Padding(
           padding: EdgeInsets.symmetric(
-              horizontal: AppSpacings.horizontalPadding.toDouble()),
+              horizontal: AppSpacings.horizontalPadding),
           child: Form(
             key: _formKey,
             child: ListView(
