+++ jim_hats_mobile/lib/presentation/views/server_down/server_down_alert_page.dart
@@ -11,7 +11,7 @@ class ServerDownAlertPage extends StatelessWidget {
       backgroundColor: Theme.of(context).colorScheme.error,
       body: Padding(
         padding: EdgeInsets.symmetric(
-          horizontal: AppSpacings.horizontalPadding.toDouble()
+          horizontal: AppSpacings.horizontalPadding
         ),
         child: Center(
           child: Column(
